:sparkles: Added new Kubernetes applications
Two new Kubernetes applications have been added. The first application is configured to automatically sync with the source repository and self-heal if necessary. It also has a sync option to create its own namespace. The second application shares similar configurations but is specifically set up in the 'argocd' namespace. Both applications are part of the default project and target the HEAD revision of their respective paths in the source repository.