:sparkles: Added Octobot application and deployment
This commit introduces a new Application for Octobot in the ArgoCD namespace. The application is configured to automatically sync from the HEAD of the main repository, with pruning and self-healing enabled.

Additionally, a Deployment for Octobot has been added. This includes configuration for persistent volume claims, container setup, and various policies such as termination grace period and restart policy. The deployment strategy is set to 'Recreate' with a revision history limit of 10.