:sparkles: Added IngressRoute and Certificate for open-webui
This commit introduces a new IngressRoute and Certificate for the 'open-webui' service. The IngressRoute is configured to match the host 'ai.mizar.scalar.cloud' with priority 10, and it points to the 'open-webui' service on port 80. TLS is enabled using a secret named 'open-webui-tls'. Additionally, a new Certificate has been created in the same namespace with DNS names set to "ai.mizar.scalar.cloud" issued by 'le-staging'.