:sparkles: Added Ingress and Service for epic-free-games
In this update, we've introduced an IngressRoute and a Service for the 'epic-free-games' application. The IngressRoute is configured with specific entry points, routes, middlewares, services and TLS settings. A Certificate has also been added to ensure secure communication.

The new Service is set up with specific ports, selectors, type of service and traffic policies. This will help in managing network access within the Kubernetes cluster.