:sparkles: Added Ingress and Service for free-games-claimer
In this update, we've introduced an IngressRoute and a Service for the free-games-claimer. The IngressRoute is configured with specific entry points, routes, middlewares, services, and TLS settings. A Certificate has also been added to ensure secure communication.

The new Service is set up with specific ports, selectors, type of service, session affinity settings as well as IP family policies. This will help in managing network traffic within the cluster effectively.