:sparkles: Added Ingress and Service for FoundryVTT
Introduced new IngressRoute and Certificate configurations under the 'foundry' namespace. The IngressRoute is set to match a specific host with priority, and it's linked to a service on the HTTP port. The Certificate configuration includes DNS names and references a ClusterIssuer.

Also added a new Service configuration in the 'foundry' namespace. This service exposes an HTTP port, uses TCP protocol, and selects instances based on app labels. It's configured as ClusterIP type with no session affinity, IPv4 family policy, and cluster internal traffic policy.