:sparkles: Added Ingress and Service for Glance
Introduced a new IngressRoute and Certificate for the Glance service in the 'glance' namespace. The IngressRoute is configured to match a specific host with priority, and it's linked to the Glance service on HTTP port. A TLS secret is also specified.

Additionally, created a new Service for Glance in the 'glance' namespace. This service exposes an HTTP port and uses TCP protocol. It selects pods based on certain labels, has no session affinity, supports only IPv4 traffic, and allows cluster-wide internal traffic.