:sparkles: Added Ingress and Service for Octobot
Introduced new IngressRoute and Certificate configurations in the ingress.yaml file. The IngressRoute is set to match a specific host with priority, includes middleware, and points to a service within the same namespace. The Certificate configuration specifies a secretName, dnsNames, and an issuerRef.

Also added a new Service configuration in the service.yaml file. This service exposes port 80 (targeting port 5001) as ClusterIP type with no session affinity. It also defines IP family policy and internal traffic policy settings.