:sparkles: Added Ingress and Service for Ollama
Introduced a new IngressRoute and Certificate in the ollama namespace, configured to use websecure entry point with home-networks middleware. The route matches a specific host and directs traffic to the newly created ollama service on port http.

Also added a new ClusterIP type Service in the ollama namespace. This service listens on port 80 (http) and routes traffic to targets labeled with 'ollama'. It's configured for IPv4 only and allows cluster-wide internal traffic.