:rocket: Added ArgoCD Application and IngressRoute
Added a new ArgoCD Application for the kube-system namespace. The application is configured to sync from a specific GitHub repository, with automated pruning and self-healing enabled.

Also added an IngressRoute for the hubble service in the kube-system namespace. This includes a rule matching a specific host, middleware reference, and associated service details. A Certificate resource has also been defined for TLS encryption using cert-manager.