:sparkles: Added Plex application to ArgoCD
This commit introduces the Plex media server application into our ArgoCD setup. The new files include a definition for the Plex application in ArgoCD, an IngressRoute for Traefik, a Service and StatefulSet for Kubernetes, and a Certificate for cert-manager. The configuration includes automated sync policy with pruning and self-healing enabled. The service is set up as a ClusterIP type with TCP protocol on port 32400. The statefulset uses the latest plexinc/pms-docker image from Docker Hub and has been configured with persistent volume claims for config and media data, as well as an emptyDir volume for transcoding data.