:sparkles: Added Authentik Application and IngressRoute
Introduced a new application configuration for Authentik in the ArgoCD namespace. The application is set to sync from a specific GitHub repository with automated pruning and self-healing enabled.

Also added an IngressRoute for the Authentik server in the authentik namespace, configured with websecure entry point and TLS certificate details.

Lastly, created a new service for the Authentik server within the authentik namespace, defining both HTTP and HTTPS ports along with their target ports.