:sparkles: Added new Kubernetes configurations
This update introduces several new Kubernetes configuration files. These include the creation of a new application in ArgoCD, deployments for Redis and Searxng, services for both applications, and persistent volume claims. The sync policy is set to automated with prune and self-heal enabled. The deployment strategy for both applications is set to 'Recreate'.