:sparkles: Added Ingress to Searxng
In this update, we've introduced an Ingress resource for the Searxng service in Kubernetes. This new addition will help manage external access to our services running within the cluster. The ingress is configured with a default backend pointing to the Searxng service on HTTP port. We've also updated the kustomization.yaml file to include this new ingress.yaml resource.