:sparkles: Added Searxng-Redis Application
A new Kubernetes application for Searxng-Redis has been added. This application is configured to use the Redis chart from Bitnami's Helm repository, with authentication disabled. The sync policy is set to automated with both prune and self-heal options enabled. A new namespace will be created during synchronization if it doesn't exist already.