:sparkles: Added new Application in ArgoCD
A new Application named 'coroot' has been added to the ArgoCD configuration. This application is set to deploy into a namespace of the same name on the default Kubernetes server. The source for this application is a Helm chart from coroot's GitHub repository, specifically version 0.3.0 of the 'coroot-ce' chart.

The sync policy for this application is automated with both prune and selfHeal options enabled, and it also includes an option to create its own namespace if it doesn't exist already. The helm values have been configured to use 2 shards and 2 replicas for clickhouse.