:sparkles: Added Traefik application and ingress
This commit introduces the Traefik application configuration in ArgoCD and its corresponding Kubernetes Ingress. The Application is set to automatically sync with the Mizar repository's HEAD revision, and a new namespace will be created if it doesn't exist. The Ingress is configured to route traffic to the Traefik dashboard service on the 'dashboard' port using Tailscale as the ingress class.