:gear: Update service configuration in YAML
The services.yaml file has been updated to replace the 'My Second Group' with a new 'Global' group. This includes the addition of a NextDNS service, complete with href, description and widget details. The widget type is set as nextdns and it uses two environment variables for profile and key.