:sparkles: Added new Application for Tautulli
A new Kubernetes application has been added for Tautulli. The application is configured to deploy in the 'plex' namespace and uses the latest image from linuxserver/tautulli. It also includes environment settings, persistence configuration, and ingress setup with TLS support. Automated sync policy with prune and self-heal options are enabled as well.