Adds security context and removes GID/UID env vars
Adds security context to the pod specification to explicitly define the user and group IDs under which the container should run.
This enhances security by ensuring that the container operates with the intended privileges.

Removes the FOUNDRY_GID and FOUNDRY_UID environment variables from the deployment configuration.
These variables are no longer necessary as the security context now manages the user and group IDs.